refactor `Search` to avoid `forwardRef`
