+++ site/src/pages/WorkspacesPage/WorkspacesButton.tsx
@@ -11,7 +11,7 @@ import {
 import type { Template } from "api/typesGenerated";
 import { Avatar } from "components/Avatar/Avatar";
 import { Loader } from "components/Loader/Loader";
-import { SearchEmpty, searchStyles } from "components/Menu/Search";
+import { SearchEmpty, searchStyles } from "components/Search/Search";
 import { OverflowY } from "components/OverflowY/OverflowY";
 import {
   Popover,
